abstract notification bell and container into own components; integrate display state into notifManager reducer
