+++ webapp/src/store/notifManager.js
@@ -6,9 +6,21 @@ const CLEARALL = 'notif/CLEARALL';
 
 const SOCKET = 'notif/SOCKET';
 
+const SHOW = 'notif/SHOW';
+
+const HIDE = 'notif/HIDE';
+
+const REPLY = 'notif/REPLY';
+
 const mountNotifs = notifications => ({ type: NOTIFS, notifications });
 
 export const SetNotifSocket = socket => ({ type: SOCKET, socket });
+
+export const ShowNotifs = () => ({ type: SHOW });
+
+export const HideNotifs = () => ({ type: HIDE });
+
+export const QuickReply = () => ({ type: REPLY });
 
 export const ClearNotif = conversationId => async () => {
   const { data } = await csrfetch(`/api/users/me/notifications/${conversationId}`, {
@@ -24,7 +36,7 @@ export const GetNotifications = () => async dispatch => {
 };
 
 export default function reducer (
-  state = { notifications: [], socket: null },
+  state = { notifications: [], socket: null, display: false, reply: false },
   { type, notifications, socket }
 ) {
   switch (type) {
@@ -34,6 +46,12 @@ export default function reducer (
       return { ...state, notifications: [] };
     case SOCKET:
       return { ...state, socket };
+    case SHOW:
+      return { ...state, display: true };
+    case HIDE:
+      return { ...state, display: false, displayQuickReply: false };
+    case REPLY:
+      return { ...state, reply: !state.reply };
     default:
       return state;
   }
